requirements.txt
streamlit
pandas
supabase
gspread
google-auth
python-dotenv
Pillow